Climbing: Refactor adding website
